Only show toast on successful login or account creation
