+++ packages/template-retail-react-app/app/pages/login/index.jsx
@@ -26,6 +26,8 @@ import useEinstein from '@salesforce/retail-react-app/app/hooks/use-einstein'
 import useDataCloud from '@salesforce/retail-react-app/app/hooks/use-datacloud'
 import LoginForm from '@salesforce/retail-react-app/app/components/login'
 import PasswordlessEmailConfirmation from '@salesforce/retail-react-app/app/components/email-confirmation/index'
+import PasskeyRegistrationModal from '@salesforce/retail-react-app/app/components/passkey-registration-modal'
+import {usePasskeyRegistration} from '@salesforce/retail-react-app/app/hooks/use-passkey-registration'
 import {
     API_ERROR_MESSAGE,
     INVALID_TOKEN_ERROR,
@@ -59,7 +61,8 @@ const Login = ({initialView = LOGIN_VIEW}) => {
     const login = useAuthHelper(AuthHelpers.LoginRegisteredUserB2C)
     const loginPasswordless = useAuthHelper(AuthHelpers.LoginPasswordlessUser)
     const authorizePasswordlessLogin = useAuthHelper(AuthHelpers.AuthorizePasswordless)
-    const {passwordless = {}, social = {}} = getConfig().app.login || {}
+    const config = getConfig()
+    const {passwordless = {}, social = {}, passkey = {}} = config.app.login || {}
     const isPasswordlessEnabled = !!passwordless?.enabled
     const isSocialEnabled = !!social?.enabled
     const idps = social?.idps
@@ -74,6 +77,7 @@ const Login = ({initialView = LOGIN_VIEW}) => {
     const [currentView, setCurrentView] = useState(initialView)
     const [passwordlessLoginEmail, setPasswordlessLoginEmail] = useState('')
     const [redirectPath, setRedirectPath] = useState('')
+    const {showToast, passkeyModal} = usePasskeyRegistration()
 
     const handleMergeBasket = () => {
         const hasBasketItem = baskets?.baskets?.[0]?.productItems?.length > 0
@@ -172,11 +176,40 @@ const Login = ({initialView = LOGIN_VIEW}) => {
 
     // If customer is registered push to account page and merge the basket
     useEffect(() => {
-        if (isRegistered) {
-            handleMergeBasket()
-            const redirectTo = redirectPath ? redirectPath : '/account'
-            navigate(redirectTo)
+        if (!isRegistered) {
+            return
         }
+
+        handleMergeBasket()
+        const redirectTo = redirectPath ? redirectPath : '/account'
+
+        if (passkey?.enabled) {
+            // Show passkey registration modal only if Webauthn feature flag is enabled and compatible with the browser
+            if (
+                window.PublicKeyCredential &&
+                // eslint-disable-next-line no-undef
+                PublicKeyCredential.isUserVerifyingPlatformAuthenticatorAvailable &&
+                // eslint-disable-next-line no-undef
+                PublicKeyCredential.isConditionalMediationAvailable
+            ) {
+                Promise.all([
+                    // eslint-disable-next-line no-undef
+                    PublicKeyCredential.isUserVerifyingPlatformAuthenticatorAvailable(),
+                    // eslint-disable-next-line no-undef
+                    PublicKeyCredential.isConditionalMediationAvailable()
+                ]).then((results) => {  
+                    if (results.every((r) => r === true)) {
+                        showToast()
+                    }
+                    // Navigate after passkey check completes (whether toast is shown or not)
+                    navigate(redirectTo)
+                })
+                return
+            }
+        }
+
+        // Navigate immediately if passkey is not enabled or not available
+        navigate(redirectTo)
     }, [isRegistered, redirectPath])
 
     /**************** Einstein ****************/
@@ -186,6 +219,7 @@ const Login = ({initialView = LOGIN_VIEW}) => {
     }, [])
 
     return (
+        <>
         <Box data-testid="login-page" bg="gray.50" py={[8, 16]}>
             <Heading as="h1" srOnly>
                 <FormattedMessage defaultMessage="Sign In" id="login.title.sign_in" />
@@ -224,6 +258,8 @@ const Login = ({initialView = LOGIN_VIEW}) => {
                 )}
             </Container>
         </Box>
+        <PasskeyRegistrationModal isOpen={passkeyModal.isOpen} onClose={passkeyModal.onClose} />
+        </>
     )
 }
